feat(PT-51): extract BookingCategory component, apply styles, and refactor code
